+++ src/Components/Footer.jsx
@@ -6,22 +6,8 @@ export default function Footer() {
         <footer className="dark:text-white/50 text-black/50 text-xs text-center mt-20 space-y-1">
 
 
-            <a href="https://next--similarify.netlify.app/">
-                Try out new features ✨
-            </a>
-
-            <p>
-                Found a bug or want a feature? Submit an issue here <a href="https://github.com/nabilridhwan/similarify/issues">here</a>
-            </p>
-
             <p className="flex justify-center items-center">
-                Created with <FaHeart className="mx-1" /> by <a href="https://nabilridhwan.github.io" className="ml-1">Nabil Ridhwan </a> &copy; {new Date().getFullYear()}
-            </p>
-
-
-
-            <p>
-                <a href="https://github.com/nabilridhwan/Similarify">Similarify is open source!</a>
+                Created with <FaHeart className="mx-1" /> by <a href="https://nabilridhwan.github.io" className="ml-1">Nabil Ridhwan</a>&nbsp;&copy; {new Date().getFullYear()}
             </p>
 
             <p>
@@ -30,10 +16,10 @@ export default function Footer() {
 
 
 
-            {/* <p>
-                <Link to="/about">About</Link>
+            <p>
+                <Link to="/about">About / FAQ</Link>
             </p>
- */}
+
 
         </footer>
     )